Reformat source
