+++ keyboards/asmodus/plaquead/info.json
@@ -1,90 +1,16 @@
  {
-    "keyboard_name": "Plaque A.D RGB",
     "maintainer": "Mai The San",
     "manufacturer": "Asmodus",
     "processor": "RP2040",
     "bootloader": "rp2040",
-    "usb": {
-        "vid": "0x5A50",
-        "pid": "0x1875",
-        "device_version": "0.0.1"
-    },
-    "features": {
-        "bootmagic": true,
-        "mousekey": true,
-        "extrakey": true,
-        "console": false,
-        "command": false,
-        "nkro": true,
-        "backlight": false,
-        "rgblight": false,
-        "audio": false,
-        "key_lock": false,
-        "encoder": false,
-        "rgb_matrix": true
-    },
     "diode_direction": "COL2ROW",
     "matrix_pins": {
         "rows": ["GP3", "GP1", "GP0", "GP25", "GP11", "GP9"],
         "cols": ["GP29", "GP27", "GP26", "GP24", "GP23", "GP22", "GP21", "GP20", "GP19", "GP18", "GP17", "GP16", "GP15", "GP14", "GP13", "GP5", "GP4"]
     },
-    "ws2812": {
-        "pin": "GP2",
-        "driver": "vendor"
+    "dynamic_keymap": {
+        "layer_count": 6
     },
-    "rgb_matrix": {
-        "animations": {
-            "alphas_mods": true,
-            "gradient_up_down": true,
-            "gradient_left_right": true,
-            "breathing": true,
-            "band_sat": true,
-            "band_val": true,
-            "band_pinwheel_sat": true,
-            "band_pinwheel_val": true,
-            "band_spiral_sat": true,
-            "band_spiral_val": true,
-            "cycle_all": true,
-            "cycle_left_right": true,
-            "cycle_up_down": true,
-            "rainbow_moving_chevron": true,
-            "cycle_out_in": true,
-            "cycle_out_in_dual": true,
-            "cycle_pinwheel": true,
-            "cycle_spiral": true,
-            "dual_beacon": true,
-            "rainbow_beacon": true,
-            "rainbow_pinwheels": true,
-            "raindrops": true,
-            "jellybean_raindrops": true,
-            "hue_breathing": true,
-            "hue_pendulum": true,
-            "hue_wave": true,
-            "fractal": true,
-            "pixel_rain": true,
-            "pixel_flow": true,
-            "typing_heatmap": true,
-            "digital_rain": true,
-            "solid_reactive_simple": true,
-            "solid_reactive": true,
-            "solid_reactive_wide": true,
-            "solid_reactive_multiwide": true,
-            "solid_reactive_cross": true,
-            "solid_reactive_multicross": true,
-            "solid_reactive_nexus": true,
-            "solid_reactive_multinexus": true,
-            "splash": true,
-            "multisplash": true,
-            "solid_splash": true,
-            "solid_multisplash": true
-        },
-        "default": {
-            "val": 60
-        },
-        "driver": "ws2812",
-        "sleep": true
-    },
-
     "url": "https://paramountkeeb.com/products/plaque-a-d",
     "layouts": {
         "LAYOUT": {
