+++ keyboards/asmodus/plaquead/readme.md
@@ -1,6 +1,6 @@
 # Plaque A.D
 
-![Plaque A.D](https://i.imgur.com/axShuI3h.jpeg)
+![Plaque A.D](https://i.imgur.com/axShuI3.jpeg)
 
 A Hotswap Multi-layout F13 TKL PCB (h88c compatible) powered by RP2040.
 
@@ -11,12 +11,12 @@ A Hotswap Multi-layout F13 TKL PCB (h88c compatible) powered by RP2040.
 Make example for this keyboard (after setting up your build environment):
 
     make asmodus/plaquead:default
-    make asmodus/plaqueadrgb:default
+    make asmodus/plaquead/rgb:default
 
 Flashing example for this keyboard:
 
     make asmodus/plaquead:default:flash
-    make asmodus/plaqueadrgb:default:flash
+    make asmodus/plaquead/rgb:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
