+++ keyboards/asmodus/plaquead/rgb/rgb.c
@@ -34,19 +34,17 @@ led_config_t g_led_config = { {
     { 83       , 84         , 85         , NO_LED     , NO_LED     , NO_LED     , 86         , NO_LED     , NO_LED     , NO_LED     , 88         , 89         , 90         , 91         , 92         , 94         , 95        }
 },{
 // "Fine-tuned" complex configuration
-{255,0  }, {240,0  }, {225,0  }, {207,0  }, {188,0  }, {174,0  }, {163,0  }, {159,0  }, {144,0  }, {126,0  }, {111,0  }, { 96,0  }, { 81,0  }, { 63,0  },            { 48,0  }, { 33,0  }, { 18,0  }, {  0,0  }, 
-{  0,15 }, { 15,15 }, { 30,15 }, { 44,15 }, { 59,15 }, { 74,15 }, { 89,15 }, {103,15 }, {118,15 }, {133,15 }, {148,15 }, {163,15 }, {177,15 }, {192,15 }, {199,15 }, {207,15 }, {225,15 }, {240,15 }, {255,15 }, 
-{255,27 }, {240,27 }, {225,27 }, {203,27 }, {185,27 }, {170,27 }, {155,27 }, {140,27 }, {126,27 }, {111,27 }, { 96,27 }, { 81,27 }, { 67,27 }, { 52,27 },                       { 37,27 }, { 22,27 }, {  4,27 }, 
-{  6,40 }, { 26,40 }, { 41,40 }, { 55,40 }, { 70,40 }, { 85,40 }, {100,40 }, {115,40 }, {129,40 }, {144,40 }, {159,40 }, {174,40 }, {198,40 }, 
-{207,52 }, {187,52 }, {194,52 }, {166,52 }, {152,52 }, {137,52 }, {122,52 }, {107,52 }, { 92,52 }, { 78,52 }, { 63,52 }, { 48,52 }, { 33,52 }, { 18,52 }, {  9,52 }, {  2,52 },
+{224,0  }, {211,0  }, {198,0  }, {182,0  }, {166,0  }, {153,0  }, {148,0  }, {140,0  }, {127,0  }, {110,0  }, { 97,0  }, { 84,0  }, { 71,0  }, { 55,0  },            { 42,0  }, { 29,0  }, { 16,0  }, {  0,0  }, 
+{  0,15 }, { 13,15 }, { 26,15 }, { 39,15 }, { 52,15 }, { 65,15 }, { 78,15 }, { 91,15 }, {104,15 }, {117,15 }, {130,15 }, {143,15 }, {156,15 }, {175,15 }, {169,15 }, {182,15 }, {198,15 }, {211,15 }, {224,15 },
+{224,27 }, {211,27 }, {198,27 }, {179,27 }, {162,27 }, {149,27 }, {136,27 }, {123,27 }, {110,27 }, { 97,27 }, { 84,27 }, { 71,27 }, { 58,27 }, { 45,27 },                       { 32,27 }, { 19,27 }, {  3,27 }, 
+{  5,40 }, { 23,40 }, { 36,40 }, { 49,40 }, { 62,40 }, { 75,40 }, { 88,40 }, {101,40 }, {114,40 }, {127,40 }, {140,40 }, {153,40 }, {174,40 }, 
+{182,52 }, {170,52 }, {164,52 }, {146,52 }, {133,52 }, {120,52 }, {107,52 }, { 94,52 }, { 81,52 }, { 68,52 }, { 55,52 }, { 42,52 }, { 29,52 }, { 16,52 }, {  8,52 }, {  2,52 }, 
+{  5,64 }, { 21,64 }, { 37,64 },                                  { 83,64 }, { 91,64 },                       {131,64 }, {151,64 }, {167,64 }, {183,64 },            {198,64 }, {211,52 }, {211,64 }, {224,64 }, 
 
-{  6,64 }, { 24,64 }, { 43,64 },                                  { 94,64 }, {109,64 },                       {150,64 }, {170,64 }, {189,64 }, {207,64 },            {225,64 }, {  240,52 }, {240,64 }, {255,64 }, 
-
-
-{255,57 }, {255,45 }, {255,32 }, {255,20 }, 
-{255,5  }, {237,5  }, {220,5  }, {203,5  }, {186,5  }, {169,5  }, {152,5  }, {135,5  }, {118,5  }, {101,5  }, { 86,5  }, { 68,5  }, { 51,5  }, { 34,5  }, { 17,5  }, 
+{224,57 }, {224,45 }, {224,32 }, {224,20 }, 
+{224,5  }, {209,5  }, {194,5  }, {179,5  }, {164,5  }, {149,5  }, {134,5  }, {119,5  }, {104,5  }, {89,5  }, { 74,5  }, { 59,5  }, { 44,5  }, { 29,5  }, { 14,5  }, 
 {  0,5  }, {  0,20 }, {  0,32 }, {  0,45 }, 
-{  0,57 }, { 17,57 }, { 34,57 }, { 51,57 }, { 68,57 }, { 85,57 }, {101,57 }, {118,57 }, {135,57 }, {152,57 }, {169,57 }, {186,57 }, {203,57 }, {220,57 }, {237,57 },
+{  0,57 }, { 14,57 }, { 29,57 }, { 44,57 }, { 59,57 }, { 74,57 }, { 89,57 }, {104,57 }, {119,57 }, {134,57 }, {149,57 }, {164,57 }, {179,57 }, {194,57 }, {209,57 },
