+++ keyboards/asmodus/plaquead/rules.mk
@@ -1,0 +1,1 @@
+DEFAULT_FOLDER = asmodus/plaquead/standard
