+++ keyboards/asmodus/plaquead/standard/config.h
@@ -14,12 +14,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-#define DYNAMIC_KEYMAP_LAYER_COUNT 6
-
 #define RP2040_BOOTLOADER_DOUBLE_TAP_RESET
 #define RP2040_BOOTLOADER_DOUBLE_TAP_RESET_TIMEOUT 500U
 
 #define RP2040_FLASH_GENERIC_03H
-
-#define LED_SCROLL_LOCK_PIN GP2
-#define LED_CAPS_LOCK_PIN GP28
